chore(ST-58): move signup api from user to auth module
