+++ src/users/users.controller.ts
@@ -1,14 +1,12 @@
-import { Body, Controller, Get, Patch, Post, UseGuards, UseInterceptors } from "@nestjs/common";
+import { Body, Controller, Get, Patch, UseGuards, UseInterceptors } from "@nestjs/common";
 
 import { makeTokenizedUser } from "@/auth/auth.helpers";
 import { ITokenizedUser } from "@/auth/auth.interfaces";
-import { AuthService } from "@/auth/auth.service";
 import { CurrentUser } from "@/auth/decorators/current-user.decorator";
 import { JwtAuthGuard } from "@/auth/guards/jwt-auth.guard";
-import { UniqueCodeGuard } from "@/auth/guards/unique-code.guard";
 import { ResponseTransformInterceptor } from "@/common/interceptors/response-transform.interceptor";
 
-import { CreateUserDto, ResetPasswordDto, UpdateUserDto, UpdatePasswordDto } from "./users.dtos";
+import { ResetPasswordDto, UpdateUserDto, UpdatePasswordDto } from "./users.dtos";
 import { UsersSerializer } from "./users.serializer";
 import { UsersService } from "./users.service";
 
@@ -17,21 +15,8 @@ import { UsersService } from "./users.service";
 export class UsersController {
   constructor(
     private readonly usersService: UsersService,
-    private readonly authService: AuthService,
     private readonly usersSerializer: UsersSerializer,
   ) {}
-  @UseGuards(UniqueCodeGuard)
-  @Post("signup")
-  async registerUser(@Body() createUserDto: CreateUserDto) {
-    const newUser = await this.usersService.createUser(createUserDto);
-
-    const accessToken = await this.authService.createAccessToken(newUser);
-
-    return {
-      accessToken,
-      user: makeTokenizedUser(newUser),
-    };
-  }
 
   @Get("me")
   @UseGuards(JwtAuthGuard)
